+++ src/common/index.ts
@@ -1,6 +1,6 @@
-import type * as t from '../types';
 import * as Util from './util';
 
 export * from './const.hashes';
 export * from './libs';
-export { Util, type t };
+export { type t } from './types.t';
+export { Util };
