+++ src/common/types.t.ts
@@ -1,0 +1,2 @@
+import type * as t from '../types';
+export type { t };
